requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.110.1
inference==0.10.0
mangum==0.17.0
numpy==1.25.2
opencv-python-headless==4.8.0.74
pillow==10.3.0
supervision==0.19.0
uvicorn
python-multipart
